requirements-api.txt
# FastAPI Web API Layer Dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9  # File upload support

# Database
sqlalchemy>=2.0.0

# ASGI server (production)
gunicorn>=21.0.0

# SSE streaming (Phase 2)
sse-starlette>=2.0.0

# Testing
httpx>=0.27.0  # For async test client
pytest-asyncio>=0.23.0  # Async test support
